[DESIGN SYSTEM - ACTIONLINL] add new test
