+++ src/actionLink/_actionLink.scss
@@ -1,6 +1,6 @@
 @import "node_modules/bootstrap/scss/functions";
-@import '../../scss/variables';
-@import '../../scss/mixins/elevation';
+@import '../scss/variables';
+@import '../scss/mixins/elevation';
 
 .actionLink {
   display: flex;
