--- src/actionLinkGrid2Columns/_actionLinkGrid.scss
@@ -1,23 +1,0 @@
-@import "node_modules/bootstrap/scss/functions";
-@import '../scss/variables';
-@import '../scss/mixins/elevation';
-
-.titleContainer{
-
-  display: flex;
-  align-items: center;
-  padding:  nth(nth($spacers, 5), 2) nth(nth($spacers, 7), 2);
-  background-color: $neutral-20;
-  color: $neutral-50;
-  font-family: $font-family-sans-serif;
-  font-size: 1.125rem;
-
-  @include elevationElementCreator1();
-
-}
-
-.suffix{
-  padding-right: nth(nth($spacers, 5), 2);
-}
-
-
